kv: make txn entry size limit configurable (#18231)

+++ store/tikv/2pc.go
@@ -338,7 +338,7 @@ func (c *twoPhaseCommitter) initKeysAndMutations() error {
 		}
 		mutations.push(op, k, value, isPessimisticLock)
 		entrySize := len(k) + len(v)
-		if entrySize > kv.TxnEntrySizeLimit {
+		if uint64(entrySize) > kv.TxnEntrySizeLimit {
 			return kv.ErrEntryTooLarge.GenWithStackByArgs(kv.TxnEntrySizeLimit, entrySize)
 		}
 		size += entrySize
